Accept constraints being dropped between subclasses
Since the initial deployment, it became evident that there was a pattern
incorrectly enforced.

Say a class C specifies a certain constraint on a property P, such as
requiring a certain maximum count.  A subclass of C, C', can make a
property shape also on P that refines any number of other constraints,
and not mention of a maximum count.  (And similarly for the other
constraint dimensions.)  It would often be convenient to do so in order
to specify that C' requires P to be stricter in just one certain
dimension, and delegate the rest of the constraints to the superclass.
The overall property shape on C' is indeed more lax in a dimension than
C is, but all of the constraints in C would still apply to C', so it is
unhelpful to require C' repeat those constraints.

This patch removes all "Dropped"-constraint errors from the validation
review.  It will still be considered an error to make an explicit
expansion, such as having C require a minimum of 2 instances of P and C'
require a minimum of 1.  But now, absence of a constraint in C' will be
treated as a delegation.

Unit test base data are also updated in this patch.

Technical debt warning:  The test cases were NOT fully reviewed to move
components no longer necessary for the XFAIL tests into the PASS tests.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>